+++ RealEstate.Application/DTOs/BuildingProperty/CreateBuildingPropertyDTO.cs
@@ -1,10 +1,13 @@
 ﻿namespace RealEstate.Application.DTOs.BuildingProperty
 {
+    /// <summary>
+    /// Represents a data transfer object (DTO) for creating a building property.
+    /// </summary>
     public class CreateBuildingPropertyDTO
     {
         public string Name { get; set; }
         public string Address { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string CodeInternal { get; set; }
         public int Year { get; set; }
         public Guid OwnerId { get; set; }
